Temporary rules

+++ src/main/java/classSuggestion/repository/SubjectRepository.java
@@ -11,7 +11,7 @@ import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
-public interface SubjectRepository extends JpaRepository<Subject, Long> {
+public interface SubjectRepository extends JpaRepository<Subject, Integer> {
 
     @Query("FROM Subject s WHERE s.id NOT in:idList")
     List<Subject> findMissingSubjects(@Param("idList") List<Integer> subjectIds);
